refactor(cli): migrate and unify update-exports logic
Reorganized update-exports structure into a CLI command under `scripts/cli`. Merged utility functions, schemas, and configurations into a centralized module, simplifying package export handling. Removed redundant code and ensured compatibility with ts-morph and zod for type validation.
