+++ scripts/cli/eslint.config.js
