+++ scripts/cli/package.json
@@ -1,31 +1,38 @@
 {
-  "name": "update-exports",
+  "name": "@codefast/cli",
   "version": "0.0.1-canary.0",
   "private": true,
   "description": "",
   "license": "MIT",
   "author": "Vuong Phan <mr.thevuong@gmail.com>",
   "type": "module",
-  "main": "dist/index.js",
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/esm/index.js",
+  "types": "./dist/types/index.d.ts",
   "bin": {
-    "update-exports": "dist/index.js"
+    "codefast": "./dist/esm/index.js"
   },
   "scripts": {
+    "build": "rslib build",
     "check-types": "tsc --noEmit",
+    "dev": "rslib build --watch",
     "lint": "eslint --max-warnings 0",
     "lint:inspect": "npx @eslint/config-inspector@latest"
   },
   "dependencies": {
     "chalk": "^5.4.1",
+    "commander": "^13.1.0",
     "glob": "^11.0.1",
-    "minimist": "^1.2.8",
+    "lodash-es": "^4.17.21",
     "ts-morph": "^25.0.1",
-    "zod": "^3.24.2"
+    "zod": "4.0.0-beta.20250414T061543"
   },
   "devDependencies": {
     "@codefast/eslint-config": "workspace:*",
     "@codefast/typescript-config": "workspace:*",
-    "@types/minimist": "^1.2.5",
+    "@rslib/core": "^0.6.3",
+    "@types/commander": "^2.12.5",
+    "@types/lodash-es": "^4.17.12",
     "@types/node": "^22.14.1",
     "eslint": "^9.24.0",
     "ts-node": "^10.9.2",
