+++ scripts/cli/src/lib/logger.ts
