+++ scripts/cli/tsconfig.build.json
@@ -1,0 +1,13 @@
+{
+  "extends": "@codefast/typescript-config/library.json",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "outDir": "dist",
+    "paths": {
+      "@/*": ["./src/*"],
+      "@codefast/eslint-config": ["./dist/*"]
+    }
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "dist", ".turbo", "coverage", "/**/*.test.*", "**/*.spec.*"]
+}
