+++ scripts/cli/tsconfig.json
