--- scripts/update-exports/src/types/config.ts
@@ -1,6 +1,0 @@
-export interface ProcessOptions {
-  dryRun: boolean;
-  verbose: boolean;
-  configPath?: string;
-  packageFilter?: string;
-}
